inserted successfully

+++ app/src/main/java/com/example/infinitymoneymanager/Meta.kt
@@ -3,13 +3,13 @@ package com.example.infinitymoneymanager
 import java.sql.Date
 class Meta(private var id: Int,
            private var nome: String,
-           private var valorAlvo: Float,
-           private var valorArrecadado: Float,
+           private var valorAlvo: Double,
+           private var valorArrecadado: Double,
            private var prazo: Date
 ) {
     fun getId(): Int {return id}
     fun getNome(): String {return nome}
-    fun getValorAlvo(): Float {return valorAlvo}
-    fun getValorArrecadado(): Float {return valorArrecadado}
+    fun getValorAlvo(): Double {return valorAlvo}
+    fun getValorArrecadado(): Double {return valorArrecadado}
     fun getPrazo(): Date {return prazo}
 }
